% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glmProvision.R
\name{glmProvision}
\alias{glmProvision}
\title{Claims reserving using glm modeling.}
\usage{
glmProvision(lossData, peMethod = "formula",
fam = 1, link = 0, B = 1000, seed = NULL)
}
\arguments{
\item{lossData}{Matrix of incremental losses \eqn{cij},
for \eqn{i = 1,...,k} origin years (rows) and for \eqn{j = 1,...,k}
development years (columns); filled with \code{NAs} for \eqn{i + j > k}.}

\item{peMethod}{Method to be used, can be \code{formula} or \code{bootstrap}. Defaults to "formula".}

\item{fam}{Index of power variance function as defined in \code{tweedie}. Defaults to Poisson.}

\item{link}{Index of power link function as defined in \code{tweedie}. Defaults to logarithmic link.}

\item{B}{Number of iterations to perform in the bootstrapping procedure. Defaults to 1000.}

\item{seed}{Seed to make bootstrap reproducible.}
}
\value{
A list of 5 elements with:
\itemize{
  \item \code{triangle} Input data (lossData).
  \item \code{glm.triangle} Fitted values by \code{glm} modeling.
  \item \code{model} \code{glm} model.
  \item \code{bootstrap.losses} (Only with \code{peMethod} = "bootstrap") Three-dimensional
   array of lower triangles from bootstrap procedure.
  \item \code{summary} Summary table.
  \item \code{params} List of output parameters.
}
}
\description{
Claims reserving using glm modeling.
}
\examples{
data("TaylorData")
glmProvision(lossData = TaylorData$lossData)
}
