% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PDR.R
\name{PDR}
\alias{PDR}
\title{Claim development results.}
\usage{
PDR(object, probs = c(0.75, 0.95, 0.995))
}
\arguments{
\item{object}{Output from \code{separateProvision} or \code{glmProvision} functions.}

\item{probs}{Quantiles of the CDR distribution.}
}
\value{
A list of 6 elements with:
\itemize{
  \item \code{meanCDRi} CDR per origin year.
  \item \code{sdCDRi} Standard deviation of CDR per origin year.
  \item \code{meanCDRT} CDR.
  \item \code{sdCDRT} Standard deviation of CDR.
  \item \code{qCDRTi} Quantiles of CDR distribution.
  \item \code{summary} Summary table.
}
}
\description{
Calculate claim development results.
}
\examples{
data("TaylorData")
res <- separateProvision(TaylorData$lossData, TaylorData$freqData)
PDR(res)
}
