% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_sepprov.R
\name{plot_sepprov}
\alias{plot_sepprov}
\title{Plots for separateProvision objects}
\usage{
plot_sepprov(object, ...)
}
\arguments{
\item{object}{separateProvision object to plot}

\item{...}{other arguments ignored (for compatibility with generic)}
}
\description{
Displays plots of a separateProvision object
}
