% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/separateProvision.R
\name{separateProvision}
\alias{separateProvision}
\title{Calculate provisions using separation methods of Taylor.}
\arguments{
\item{lossData}{Matrix of incremental losses \eqn{Cij},
for \eqn{i = 1,...,t} origin years (rows) and for \eqn{j = 1,...,t}
development years (columns); filled with \code{NAs} for \eqn{i + j > t}.}

\item{freqData}{Vector with the claim numbers by origin year \eqn{i}
for \eqn{i = 1,...t}.}

\item{modelSep}{Model to be used, can be \code{arithmetic} or \code{geometric}.}

\item{lambdaK}{Percentage of the trend in the inflation index.}

\item{B}{Number of iterations to perform in the bootstrapping procedure.}

\item{seed}{Seed to make bootstrap reproducible.}
}
\value{
A list of 5 elements with:
\itemize{
  \item \code{triangle} Input data (lossData).
  \item \code{glm.triangle} Fitted values by \code{glm} modeling.
  \item \code{model} \code{glm} model.
  \item \code{glm.triangle.bootstrap} Three-dimensional array of lower triangles from bootstrap procedure.
  \item \code{summary} Summary table.
  \item \code{params} List of output parameters.
}
}
\description{
Calculate provisions using separation methods of Taylor.
}
\examples{
data("TaylorData")
separateProvision(lossData = TaylorData$lossData,
                  freqData = TaylorData$freqData)
}
