% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TaylorData.R
\docType{data}
\name{TaylorData}
\alias{TaylorData}
\title{Taylor and Ashe data}
\format{A list with 2 elements containing example data for the package.
\itemize{
  \item \code{lossData} Matrix of incremental losses.
  \item \code{freqData} Vector with the claim numbers by origin year.
}}
\usage{
data("TaylorData")
}
\description{
Taylor and Ashe example data
}
\references{
Taylor, G. and Ashe, F., 1983. Second moments of estimates of outstanding claims. Journal of Econometrics 23, 37􏰀61.
}
\keyword{datasets}
